# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Database service API",
    "version": "0.0.1",
    "description": "This is a REST API for the database service.",
    "license": {
      "name": "Proprietary"
    },
    "contact": {
      "name": "Tomáš Kebrle"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "Returns a list of users.",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "The list of users.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "enum": [
                        "required"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Creates a new user.",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "The user's name.",
                    "example": "Tomáš Kebrle"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "fieldErrors": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "required"
                            ]
                          }
                        }
                      },
                      "example": {
                        "name": [
                          "required"
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Returns a user by ID.",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "The user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Id is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "enum": [
                        "not-valid"
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "string",
                      "enum": [
                        "not-found"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The user's name."
          },
          "id": {
            "type": "string",
            "description": "The user's id."
          }
        },
        "example": {
          "name": "John Doe",
          "id": "f7cabd53-49e1-4c93-b59e-6035811b134d"
        }
      },
      "ErrorCodes": {
        "type": "string",
        "enum": [
          "required",
          "not-found"
        ]
      }
    }
  },
  "tags": [
    {
      "name": "Users",
      "description": "API for managing users."
    }
  ]
}